時系列レイヤーを共通クラス化
